requirementstwo.txt
# Core packages
pandas==2.0.3
numpy==1.23.5
scikit-learn==1.2.2
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.13.2

# Gradient Boosting Libraries
xgboost==1.7.6
lightgbm==4.1.0
catboost==1.2.3

# ML pipeline and preprocessing
feature-engine==1.6.1
optuna==3.2.0
sweetviz==2.3.1
kneed==0.8.5

# Feature engineering
statsmodels==0.14.0
category-encoders==2.6.2

# Model explainability
shap==0.40.0
numba==0.56.4
llvmlite==0.39.1

# PostgreSQL & ORM
psycopg2-binary==2.9.7
sqlalchemy==1.4.49

# Flask API
flask==2.2.5
python-dotenv==1.0.0

# MLOps
mlflow==2.6.0
evidently

# Logging & testing
python-json-logger==2.0.7
pytest==7.4.0

# Misc
tqdm==4.67.1
boto3
redshift-connector
mlflow
python-dotenv
#pip install requests-auth-aws-sigv4


